Fixing snippet path generation
Previously, snippets (often foo/bar) could not be saved, as only $dataroot/snippets/ was created. Additionally, because the permission mask was passed as a string, the directory was created in an unreadable format.

Solution was to recursively create snippet paths, and ensure that the path is created with the correct permissions.